ref:Remove unnecessary yaml file and proper rename package.
